% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsffq_simple_functions.R
\name{fq}
\alias{fq}
\title{Frequency Factor}
\usage{
fq(F)
}
\arguments{
\item{F}{1-8, representing different frequency factor responses from the
hsffq. These can be in a dataframe, vector, or just single values}
}
\value{
a dataframe, vector, or single value of the same dimension as the
input, with each position holding the average daily servings consumed
for each food (columns) for each individual(rows).
}
\description{
The Frequency Factor function converts values 1-8, representing different
frequency factor responses from the hsffq, to average daily servings consumed
for that indevidual.
}
\examples{
test <- c(1,5,7,3,9,2,4,3,6,8)
results<- fq(test)
results

}
