% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsffq.R
\name{hsffq}
\alias{hsffq}
\title{Harvard Foood Frequency Questionnaire Nutrition Information}
\usage{
hsffq()
}
\value{
A portion size and nutrient information reference data frame.
}
\description{
This dataframe is used internally by the functions of foodquotient and
includes portion size information by age for each of the 85 foods included in
the HSFFQ from the HSFFQ user's manual. Additionally, nutrient information
is included for each of the 85 foods, pulled from the USDA's public search
tool.
}
