% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsffq_simple_functions.R
\name{micros}
\alias{micros}
\title{Micronutrients}
\usage{
micros(row)
}
\arguments{
\item{row}{contains two components.
    F1:F85 1-8, representing different frequency factor responses from the
hsffq. These will be stored in columns 2-86 in the row you plug in
.   A value representing participant's age. This will be stored in column
1 of the inputed row}
}
\value{
the row or dataframe returned will have 7 entries, representing
total daily amounts of 7 micronutrients for each participant
}
\description{
The Micronutrients function takes the age of a participant and their responses
on the hsffq to generate an estimate of the participant's total daily
micronutrients consumed for each food
}
\examples{
random_integers <- sample(1:8, 85, replace=TRUE)
vec <- c(6.2, random_integers)
micros(vec)

vec1 <- c(5.1, sample(1:8, 85, replace = TRUE))
row1 <- data.frame(t(vec1))
row2 <- data.frame(t(c(8.3, sample(1:8, 85, replace = TRUE))))
df <- rbind(row1, row2)

df_results <- data.frame()
for (i in 1:nrow(df)) {
result <- micros(df[i,])
df_results <- rbind(df_results, result)
}


}
