% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsffq_simple_functions.R
\name{quotient}
\alias{quotient}
\title{Food quotient based on hsffq results}
\usage{
quotient(row)
}
\arguments{
\item{row}{contains two components.
    F1:F85 1-8, representing different frequency factor responses from the
hsffq. These will be stored in columns 2-86 in the row you plug in
    A value representing participant's age. This will be stored in column
1 of the inputed row}
}
\value{
one value per participant will be returned, representing the food
quotient for the indevidual
}
\description{
The quotient function calculates individual level food quotients based on
the individual's answers to the hsffq. This function is only recommended to
calculate at the individual level.
}
\examples{
random_integers <- sample(1:8, 85, replace=TRUE)
vec <- c(6.2, random_integers)
quotient(vec)

vec1 <- c(5.1, sample(1:8, 85, replace = TRUE))
row1 <- data.frame(t(vec1))
row2 <- data.frame(t(c(8.3, sample(1:8, 85, replace = TRUE))))
df <- rbind(row1, row2)

df_results <- data.frame()
for (i in 1:nrow(df)) {
result <- quotient(df[i,])
df_results <- rbind(df_results, result)
}





}
