% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsffq_simple_functions.R
\docType{data}
\name{age_freq}
\alias{age_freq}
\title{Frequency Factors for CATALYST Project Children with Age of Participant}
\format{
## `age_freq`
A data frame with 32 rows and 86 columns:
\describe{
  \item{a}{age of participant}
  \item{f1}{milk frequency factor}
  \item{f2}{hot chocolate frequency factor}
  \item{f3}{cheese frequency factor}
  \item{f4}{yogurt frequency frequency factor}
  \item{f5}{ice cream frequency frequency factor}
  \item{f6}{pudding frequency factor}
  \item{f7}{orange juice frequency factor}
  \item{f8}{other juice frequency factor}
  \item{f9}{fruit drink frequency factor}
  \item{f10}{banana frequency factor}
  \item{f11}{peaches frequency factor}
  \item{f12}{mixed fruit frequency factor}
  \item{f13}{orange frequency factor}
  \item{f14}{apple and pear frequency factor}
  \item{f15}{applesauce frequency factor}
  \item{f16}{grapes frequency factor}
  \item{f17}{strawberries frequency factor}
  \item{f18}{melon frequency factor}
  \item{f19}{pineapple frequency factor}
  \item{f20}{raisins frequency factor}
  \item{f21}{corn frequency factor}
  \item{f22}{peas frequency factor}
  \item{f23}{tomato frequency factor}
  \item{f24}{peppers frequency factor}
  \item{f25}{carrot frequency factor}
  \item{f26}{broccoli frequency factor}
  \item{f27}{green beans frequency factor}
  \item{f28}{spinach frequency factor}
  \item{f29}{greens frequency factor}
  \item{f30}{mixed vegetable frequency factor}
  \item{f31}{squash frequency factor}
  \item{f32}{zucchini frequency factor}
  \item{f33}{fried potatoes frequency factor}
  \item{f34}{other potatoes frequency factor}
  \item{f35}{sweet potatoes frequency factor}
  \item{f36}{cabbage frequency factor}
  \item{f37}{lettuce frequency factor}
  \item{f38}{mayonnaise frequency factor}
  \item{f39}{chips frequency factor}
  \item{f40}{popcorn frequency factor}
  \item{f41}{crackers frequency factor}
  \item{f42}{nuts frequency factor}
  \item{f43}{cookies frequency factor}
  \item{f44}{cake frequency factor}
  \item{f45}{pie frequency factor}
  \item{f46}{jello frequency factor}
  \item{f47}{chocolate frequency factor}
  \item{f48}{candy frequency factor}
  \item{f49}{coffee frequency factor}
  \item{f50}{soda frequency factor}
  \item{f51}{sugarfree soda frequency factor}
  \item{f52}{beans frequency factor}
  \item{f53}{rice frequency factor}
  \item{f54}{pasta frequency factor}
  \item{f55}{pizza frequency factor}
  \item{f56}{tacos frequency factor}
  \item{f57}{mac and cheese frequency factor}
  \item{f58}{hot dogs frequency factor}
  \item{f59}{sausage frequency factor}
  \item{f60}{hamburger frequency factor}
  \item{f61}{tuna frequency factor}
  \item{f62}{fried fish frequency factor}
  \item{f63}{other fish frequency factor}
  \item{f64}{cold cuts frequency factor}
  \item{f65}{chicken nuggets frequency factor}
  \item{f66}{other chicken frequency factor}
  \item{f67}{pork frequency factor}
  \item{f68}{beef frequency factor}
  \item{f69}{organ meats frequency factor}
  \item{f70}{peanut butter frequency factor}
  \item{f71}{bread frequency factor}
  \item{f72}{butter frequency factor}
  \item{f73}{margarine frequency factor}
  \item{f74}{vegetabele soup frequency factor}
  \item{f75}{soup frequency factor}
  \item{f76}{tortilla frequency factor}
  \item{f77}{eggs frequency factor}
  \item{f78}{bacon frequency factor}
  \item{f79}{hot cereal frequency factor}
  \item{f80}{cold cereal frequency factor}
  \item{f81}{donuts frequency factor}
  \item{f82}{muffins frequency factor}
  \item{f83}{pancake frequency factor}
  \item{f84}{bagel frequency factor}
  \item{f85}{biscuit frequency factor}






}
}
\source{
<Baylor Human Evolutionary Biology Lab>
}
\usage{
age_freq
}
\description{
A small set of food frequency questionnaire data including 32 children living
 in the United States. f1:f85 represents the frequency with which
 participants consumed 85 respective foods.
Numbers 1-9 correspond to the following:
1: never
2: 1-3 times per month
3: once per week
4: 2-4 times per week
5: 5-6 times per week
6: 1 per day
7: 2-3 times per day
8: 4-5 times per day
9: 6 times per day
}
\keyword{datasets}
